navbar, products, single_product, and some back-end
